Move some fares classes back into sandbox
