+++ src/ext-test/java/org/opentripplanner/ext/fares/impl/GtfsFaresV2ServiceTest.java
@@ -12,13 +12,13 @@ import java.util.Map;
 import java.util.Set;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import org.opentripplanner.ext.fares.model.Distance;
+import org.opentripplanner.ext.fares.model.FareDistance;
+import org.opentripplanner.ext.fares.model.FareDistance.LinearDistance;
+import org.opentripplanner.ext.fares.model.FareLegRule;
+import org.opentripplanner.ext.fares.model.FareTransferRule;
 import org.opentripplanner.framework.geometry.SphericalDistanceLibrary;
-import org.opentripplanner.model.fare.Distance;
-import org.opentripplanner.model.fare.FareDistance;
-import org.opentripplanner.model.fare.FareDistance.LinearDistance;
-import org.opentripplanner.model.fare.FareLegRule;
 import org.opentripplanner.model.fare.FareProduct;
-import org.opentripplanner.model.fare.FareTransferRule;
 import org.opentripplanner.model.plan.Itinerary;
 import org.opentripplanner.model.plan.Place;
 import org.opentripplanner.model.plan.PlanTestConstants;
