+++ src/ext/java/org/opentripplanner/ext/fares/impl/DefaultFareServiceFactory.java
@@ -9,12 +9,12 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import org.opentripplanner.ext.fares.model.FareAttribute;
+import org.opentripplanner.ext.fares.model.FareLegRule;
 import org.opentripplanner.ext.fares.model.FareRule;
 import org.opentripplanner.ext.fares.model.FareRuleSet;
 import org.opentripplanner.ext.fares.model.FareRulesData;
+import org.opentripplanner.ext.fares.model.FareTransferRule;
 import org.opentripplanner.model.OtpTransitService;
-import org.opentripplanner.model.fare.FareLegRule;
-import org.opentripplanner.model.fare.FareTransferRule;
 import org.opentripplanner.routing.core.FareType;
 import org.opentripplanner.routing.fares.FareService;
 import org.opentripplanner.routing.fares.FareServiceFactory;
