+++ src/ext/java/org/opentripplanner/ext/fares/impl/GtfsFaresService.java
@@ -1,6 +1,8 @@
 package org.opentripplanner.ext.fares.impl;
 
+import java.util.Collection;
 import java.util.Objects;
+import org.opentripplanner.ext.fares.model.LegProducts;
 import org.opentripplanner.model.fare.ItineraryFares;
 import org.opentripplanner.model.plan.Itinerary;
 import org.opentripplanner.routing.fares.FareService;
@@ -9,12 +11,29 @@ public record GtfsFaresService(DefaultFareService faresV1, GtfsFaresV2Service fa
   implements FareService {
   @Override
   public ItineraryFares calculateFares(Itinerary itinerary) {
-    var fare = Objects.requireNonNullElse(faresV1.calculateFares(itinerary), ItineraryFares.empty());
+    var fare = Objects.requireNonNullElse(
+      faresV1.calculateFares(itinerary),
+      ItineraryFares.empty()
+    );
     var products = faresV2.getProducts(itinerary);
     fare.addItineraryProducts(products.itineraryProducts());
     if (products.itineraryProducts().isEmpty()) {
-      fare.addLegProducts(products.legProducts());
+      addLegProducts(products.legProducts(), fare);
     }
     return fare;
+  }
+  /**
+   * Add a complex set of fare products for a specific leg;
+   */
+  private static void addLegProducts(Collection<LegProducts> legProducts, ItineraryFares fares) {
+    legProducts.forEach(lp -> {
+      lp
+        .products()
+        .stream()
+        .map(LegProducts.ProductWithTransfer::product)
+        .forEach(fp -> {
+          fares.addFareProduct(lp.leg(), fp);
+        });
+    });
   }
 }
