+++ src/ext/java/org/opentripplanner/ext/fares/model/FareDistance.java
@@ -1,9 +1,6 @@
-package org.opentripplanner.model.fare;
-
-import org.opentripplanner.framework.lang.Sandbox;
+package org.opentripplanner.ext.fares.model;
 
 /** Represents a distance metric used in distance-based fare computation*/
-@Sandbox
 public sealed interface FareDistance {
   /** Represents the number of stops as a distance metric in fare computation */
   record Stops(int min, int max) implements FareDistance {}
