+++ src/ext/java/org/opentripplanner/ext/fares/model/FareLegRule.java
@@ -1,10 +1,9 @@
-package org.opentripplanner.model.fare;
+package org.opentripplanner.ext.fares.model;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import org.opentripplanner.framework.lang.Sandbox;
+import org.opentripplanner.model.fare.FareProduct;
 
-@Sandbox
 public record FareLegRule(
   @Nullable String legGroupId,
   @Nullable String networkId,
