+++ src/ext/java/org/opentripplanner/ext/fares/model/FareRulesData.java
@@ -2,8 +2,6 @@ package org.opentripplanner.ext.fares.model;
 
 import java.util.ArrayList;
 import java.util.List;
-import org.opentripplanner.model.fare.FareLegRule;
-import org.opentripplanner.model.fare.FareTransferRule;
 
 public record FareRulesData(
   List<FareAttribute> fareAttributes,
