+++ src/ext/java/org/opentripplanner/ext/fares/model/FareTransferRule.java
@@ -1,11 +1,11 @@
-package org.opentripplanner.model.fare;
+package org.opentripplanner.ext.fares.model;
 
 import java.time.Duration;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import org.opentripplanner.framework.lang.Sandbox;
+import org.opentripplanner.model.fare.FareProduct;
 
-@Sandbox
 public record FareTransferRule(
   @Nullable String fromLegGroup,
   @Nullable String toLegGroup,
