+++ src/ext/java/org/opentripplanner/ext/fares/model/LegProducts.java
@@ -1,13 +1,13 @@
-package org.opentripplanner.model.fare;
+package org.opentripplanner.ext.fares.model;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 import org.opentripplanner.framework.lang.Sandbox;
+import org.opentripplanner.model.fare.FareProduct;
 import org.opentripplanner.model.plan.Leg;
 import org.opentripplanner.model.plan.ScheduledTransitLeg;
 
-@Sandbox
 public record LegProducts(
   Leg leg,
   Optional<ScheduledTransitLeg> nextLeg,
