+++ src/test/java/org/opentripplanner/gtfs/mapping/FareTransferRuleMapperTest.java
@@ -76,7 +76,7 @@ class FareTransferRuleMapperTest {
     return fareProduct;
   }
 
-  private org.opentripplanner.model.fare.FareTransferRule map(
+  private org.opentripplanner.ext.fares.model.FareTransferRule map(
     FareProduct fareProduct,
     FareTransferRule rule
   ) {
